Add back SignIdentity message

+++ crypto/local/bip32.c
@@ -138,7 +138,7 @@ uint32_t hdnode_fingerprint(HDNode *node)
 	hdnode_fill_public_key(node);
 	sha256_Raw(node->public_key, 33, digest);
 	ripemd160(digest, 32, digest);
-	fingerprint = (digest[0] << 24) + (digest[1] << 16) + (digest[2] << 8) + digest[3];
+	fingerprint = ((uint32_t)digest[0] << 24) + (digest[1] << 16) + (digest[2] << 8) + digest[3];
 	MEMSET_BZERO(digest, sizeof(digest));
 	return fingerprint;
 }
